Add task controller class
