+++ src/controllers/board.js
@@ -1,0 +1,80 @@
+import {CARDS_PER_PAGE} from '../const';
+import {isEscPressed, render} from '../util';
+import {Board} from '../components/board';
+import {Sort} from '../components/sort';
+import {MoreBtn} from '../components/more-btn';
+import {TaskController} from './task';
+
+class BoardController {
+  constructor(container, tasks) {
+    this._container = container;
+    this._tasks = tasks;
+    this._tasksToRender = tasks;
+    this._renderedTasks = 0;
+    this._sortEl = new Sort().getElement();
+    this._boardEl = new Board(this._tasks).getElement();
+    this._tasksEl = this._boardEl.querySelector(`.board__tasks`);
+    this._loadMoreEl = new MoreBtn().getElement();
+  }
+
+  _renderTasks(isContinues = true) {
+    const quantity = (isContinues) ? CARDS_PER_PAGE : this._renderedTasks;
+    this._renderedTasks = (isContinues) ? this._renderedTasks : 0;
+    const endIndex = this._renderedTasks + quantity;
+
+    for (let i = this._renderedTasks; i < endIndex && i < this._tasksToRender.length; i++) {
+      new TaskController(this._tasksEl, this._tasksToRender[i]).init();
+      this._renderedTasks++;
+    }
+
+    if (this._renderedTasks >= this._tasksToRender.length && this._loadMoreEl) {
+      this._loadMoreEl.remove();
+    }
+  }
+
+  _onSortLinkClick(evt) {
+    evt.preventDefault();
+
+    const sortType = evt.target.dataset.type;
+
+    if (sortType) {
+      const tasksCopy = this._tasks.slice();
+      this._tasksEl.innerHTML = ``;
+
+      switch (sortType) {
+        case `default`:
+          this._tasksToRender = this._tasks;
+          this._renderTasks(false);
+          break;
+        case `date-up`:
+          const dateUpTasks = tasksCopy.sort((first, second) => first.dueDate - second.dueDate);
+
+          this._tasksToRender = dateUpTasks;
+          this._renderTasks(false);
+          break;
+        case `date-down`:
+          const dateDownTasks = tasksCopy.sort((first, second) => second.dueDate - first.dueDate);
+
+          this._tasksToRender = dateDownTasks;
+          this._renderTasks(false);
+          break;
+      }
+    }
+  }
+
+  init() {
+    this._sortEl.addEventListener(`click`, (evt) => this._onSortLinkClick(evt));
+    render(this._boardEl, this._sortEl, `begin`);
+    this._renderTasks();
+
+    this._loadMoreEl.addEventListener(`click`, (evt) => {
+      evt.preventDefault();
+      this._renderTasks();
+    });
+
+    render(this._boardEl, this._loadMoreEl);
+    render(this._container, this._boardEl);
+  }
+}
+
+export {BoardController};
